refactor: update CoverFlowView to use string view IDs
Removes viewConfigMap references in CoverFlowView components, using
string literals (coverFlow, nowPlaying) instead. Fixes circular
dependency imports.
